requirements.txt
Django
djangorestframework
django-cors-headers
psycopg
pillow
django-debug-toolbar